requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
aiosqlite
ccxt
python-dotenv
python-telegram-bot
PyYAML